requirements.txt
customtkinter~=5.2.2
pillow~=10.4.0
matplotlib~=3.9.2
pandas~=2.2.3